+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Pass fragment identifiers from navigation wrapper to book section
+
 ## 0.1.0
 
 Initial release.